% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossterm.R
\name{cor_approx_cs}
\alias{cor_approx_cs}
\title{Calculates an estimate of Neyman's bound of the cross-term of the variance of average treatment effect estimator under general regression adjsustment in randomized two-arm
experiments based on the Cauchy-Schwarz Inequality.}
\usage{
cor_approx_cs(r_obs, Z)
}
\arguments{
\item{r_obs}{The residuals of the outcome model,i.e, $Y_i(q) - \hat f_q(X_i)$.}

\item{Z}{Vector of the treatment assignment.}
}
\value{
estimate of Neyman's Cauchy-Schwarz bound on the cross-term
}
\description{
This function calculates an estimate of Neyman's bound of the cross-term of the variance of average treatment effect estimator under general regression adjsustment in randomized two-arm
experiments based on the Cauchy-Schwarz Inequality.
}
