% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossterm.R
\name{cor_approx_sharp}
\alias{cor_approx_sharp}
\title{Calculates an estimates of the cross-term of the variance of average treatment effect estimator under general regression adjsustment in randomized two-arm
experiments under extremal (comonotonic) dependence of the potential outcomes.}
\usage{
cor_approx_sharp(r_obs, Z, upper = TRUE)
}
\arguments{
\item{r_obs}{The residuals of the outcome model,i.e, $Y_i(q) - \hat f_q(X_i)$.}

\item{Z}{Vector of the treatment assignment.}
}
\value{
estimate of the cross-term of the variance average treatment effect estimator under general regression adjsustment in randomized two-arm
experiments under extremal (comonotonic) dependence of the potential outcomes.
}
\description{
This function calculates a consistent estimate of the cross-term of the variance of average treatment effect estimator under general regression adjsustment in randomized two-arm
experiments under extremal dependence of the potential outcomes
}
