% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{create_dataset}
\alias{create_dataset}
\title{Creates a simulated dataset to showcase sharp variance bounds}
\usage{
create_dataset(N = 1000, sd = 0, eps_var = 1)
}
\arguments{
\item{N}{The number of simulated units.}

\item{sd}{Simulation parameter}
}
\value{
Simulated dataset containing potential outcomes (Y0,Y1) and covariates X.
}
\description{
This function creates the simulated dataset described in (Mikhaeil, Green, 2024)
to showcase the use of and benefits of sharp variance bounds.
}
