% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceEstimators.R
\name{neyman_conventional}
\alias{neyman_conventional}
\title{Calculates (Neyman's) conventional variance estimator.}
\usage{
neyman_conventional(r_obs, Z)
}
\arguments{
\item{r_obs}{The residuals of the outcome model \eqn{Y_i(q) - \hat f_q(X_i)}.}

\item{Z}{Vector of the treatment assignment.}
}
\value{
Conventional variance estimator
}
\description{
Calculates (Neyman's) conventional variance estimator.
}
