% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceEstimators.R
\name{neyman_cs}
\alias{neyman_cs}
\title{Calculates Neyman's variance estimator based on the Cauchy-Schwarz Inequality.}
\usage{
neyman_cs(r_obs, Z)
}
\arguments{
\item{r_obs}{The residuals of the outcome model,i.e, $Y_i(q) - \hat f_q(X_i)$.}

\item{Z}{Vector of the treatment assignment.}
}
\value{
Neyman's Cauchy-Schwarz variance estimator
}
\description{
Calculates Neyman's variance estimator based on the Cauchy-Schwarz Inequality.
}
