% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{randomize}
\alias{randomize}
\title{Randomizes treatment}
\usage{
randomize(Y0, Y1, x, N_r = 0.5)
}
\arguments{
\item{Y0}{Potential outcomes under no treatment}

\item{Y1}{Potential outcomes under treatment}

\item{x}{covariates}

\item{N_r}{ratio of people receiving treatment}
}
\value{
Randomized dataset containing observed outcomes Y_obs, treatment and covariates X
}
\description{
This function simulates a CRE by randomizing treatment over a population of potential outcomes.
}
