% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{var_true}
\alias{var_true}
\title{Calculates the true variance}
\usage{
var_true(eps0, eps1)
}
\arguments{
\item{eps0}{Potential outcomes under no treatment}

\item{eps1}{Potential outcomes under treatment}
}
\value{
Randomized dataset containing observed outcomes Y_obs, treatment and covariates X
}
\description{
This function simulates a CRE by randomizing treatment over a population of potential outcomes.
}
